+++ lib/couchrest/model/property.rb
@@ -26,7 +26,7 @@ module CouchRest::Model
       if type.is_a?(Array)
         if value.nil?
           value = []
-        elsif [Hash, HashWithIndifferentAccess].include?(value.class)
+        elsif value.is_a?(Hash)
           # Assume provided as a Hash where key is index!
           data = value
           value = [ ]
@@ -39,7 +39,7 @@ module CouchRest::Model
         arr = value.collect { |data| cast_value(parent, data) }
         # allow casted_by calls to be passed up chain by wrapping in CastedArray
         CastedArray.new(arr, self, parent)
-      elsif (type == Object || type == Hash) && (value.class == Hash)
+      elsif (type == Object || type == Hash) && (value.is_a?(Hash))
         # allow casted_by calls to be passed up chain by wrapping in CastedHash
         CastedHash[value, self, parent]
       elsif !value.nil?